refactor(providers): extract shared helpers into providers/shared
(A) fix(test): add AFK_MODEL_LOCAL + sub-keys to env-bleed clear list
    in config.test.ts model-slots beforeEach/afterEach. Prevents local dev
    shells with AFK_MODEL_LOCAL/LOCAL_BASE_URL/LOCAL_API_KEY set from
    bleeding into DEFAULT_SLOT_BINDINGS assertions.

(B) extract: move three verbatim-duplicated pure functions to shared/
    - sleepWithAbort     → providers/shared/sleep-with-abort.ts
    - summarizeToolInput → providers/shared/tool-input-summary.ts
    - supportedCommands  → providers/shared/supported-commands.ts
      (collectSupportedCommands helper; both providers now delegate)
    Deleted both inline copies; updated imports in anthropic-direct/loop.ts
    and openai-compatible/query.ts.

(C) relocate: move cross-provider-imported helpers to shared/
    - auto-compact.ts       → providers/shared/auto-compact.ts
    - plan-mode-addendum.ts → providers/shared/plan-mode-addendum.ts
    - afk-mode-addendum.ts  → providers/shared/afk-mode-addendum.ts
    Original files become re-export shims so existing test imports
    (auto-compact.test.ts, plan-mode-addendum.test.ts, etc.) resolve
    without modification. openai-compatible/query.ts cross-boundary imports
    updated to ../shared/*.

(D) deferred: haiku model-ID divergence (query.ts:91 vs compact-handler.ts:54)
    is pre-existing and masked by AFK_COMPACT_MODEL env bleed in the test
    runner; reverting to avoid false signal.

Lint: tsc --noEmit clean. Tests: 1 failed (pre-existing AFK_COMPACT_MODEL
env bleed in anthropic-direct.test.ts, present on main before this branch).
